Move Response/Responses out of Http4sHelper and into common code
It's used a bunch of places in the core and in generators and is by no
means http4s-specific.
