+++ modules/codegen/src/main/scala/com/twilio/guardrail/generators/EndpointsClientGenerator.scala
@@ -2,19 +2,18 @@ package com.twilio.guardrail
 package generators
 
 import java.util.Locale
-
 import _root_.io.swagger.v3.oas.models.PathItem.HttpMethod
 import cats.arrow.FunctionK
 import cats.data.{ Ior, NonEmptyList }
 import cats.implicits._
 import com.twilio.guardrail.generators.syntax.Scala._
 import com.twilio.guardrail.generators.syntax._
+import com.twilio.guardrail.languages.ScalaLanguage
+import com.twilio.guardrail.protocol.terms.Responses
 import com.twilio.guardrail.protocol.terms.client._
 import com.twilio.guardrail.terms.RouteMeta
-import com.twilio.guardrail.languages.ScalaLanguage
 import com.twilio.guardrail.shims._
 import java.net.URI
-
 import scala.collection.JavaConverters._
 import scala.meta._
