+++ modules/codegen/src/main/scala/com/twilio/guardrail/protocol/terms/server/ServerTerm.scala
@@ -1,9 +1,10 @@
 package com.twilio.guardrail.protocol.terms.server
 
+import com.twilio.guardrail.generators.ScalaParameters
 import com.twilio.guardrail.{ RenderedRoutes, StrictProtocolElems, SupportDefinition, TracingField }
 import com.twilio.guardrail.terms.RouteMeta
 import com.twilio.guardrail.languages.LA
-import com.twilio.guardrail.generators.{ Responses, ScalaParameters }
+import com.twilio.guardrail.protocol.terms.{ Response, Responses }
 import io.swagger.v3.oas.models.Operation
 
 sealed trait ServerTerm[L <: LA, T]
